requirements.txt
cloudpickle==3.0.0
contourpy==1.2.1
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
gym-notices==0.0.8
gymnasium==0.29.1
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
setuptools==68.2.2
six==1.16.0
stable_baselines3==2.3.2
sympy==1.12
torch==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
wheel==0.41.2
